requirements.local.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
psutil
reportlab
torch
torchvision
transformers
timm
Pillow
